Finally got Somewher(Ya bois got skillz)
After a lot of research on PVectors i decided to try to translate the entire matrix whenever a turn takes place so it would look like a turn but without the animations. Made some progress today in figuring out how to change the cube status with different keycodes such as R. Right now, I  am able to randomize the cube but i need to create localized changes to emulate turns.